# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CarePath healthcare discovery, planning, appointments, demo tokens, reviews, and administration API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /hospitals:
    get:
      operationId: listHospitals
      tags: [carepath]
      summary: Search and filter demo hospital listings
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: city, in: query, schema: { type: string } }
        - { name: locality, in: query, schema: { type: string } }
        - { name: type, in: query, schema: { type: string, enum: [hospital, clinic] } }
        - { name: open, in: query, schema: { type: boolean } }
        - { name: specialty, in: query, schema: { type: string } }
        - { name: latitude, in: query, schema: { type: number, minimum: -90, maximum: 90 } }
        - { name: longitude, in: query, schema: { type: number, minimum: -180, maximum: 180 } }
      responses:
        "200":
          description: Matching hospitals
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Hospital" }
  /hospitals/{id}:
    get:
      operationId: getHospital
      tags: [carepath]
      summary: Get a hospital profile with doctors and reviews
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Hospital details
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HospitalDetail" }
        "404": { description: Hospital not found }
...
  /emergency-facilities:
    get:
      operationId: listEmergencyFacilities
      tags: [carepath]
      summary: List sample emergency facilities
      parameters:
        - { name: city, in: query, schema: { type: string } }
        - { name: category, in: query, schema: { type: string } }
      responses:
        "200":
          description: Sample emergency facilities; availability is not live
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/EmergencyFacility" }
...
  /hospitals/{id}/reviews:
    post:
      operationId: createReview
      tags: [carepath]
      summary: Submit a review for a hospital
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReviewInput" }
      responses:
        "201":
          description: Review submitted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Review" }
        "401": { description: Sign in required }
...
components:
  schemas:
...
    Hospital:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        type: { type: string, enum: [hospital, clinic] }
        address: { type: string }
        city: { type: string }
        locality: { type: string }
        contactPhone: { type: string }
        distanceKm: { type: ["number", "null"] }
        specialties: { type: array, items: { type: string } }
        doctorCount: { type: integer }
        rating: { type: number }
        reviewCount: { type: integer }
        waitingMinutes: { type: integer }
        isOpen: { type: boolean }
        isDemo: { type: boolean }
        emergency: { type: boolean }
        hours: { type: string }
      required: [id, name, type, address, city, locality, contactPhone, distanceKm, specialties, doctorCount, rating, reviewCount, waitingMinutes, isOpen, isDemo, emergency, hours]
    HospitalDetail:
      allOf:
        - { $ref: "#/components/schemas/Hospital" }
        - type: object
          properties:
            doctors: { type: array, items: { $ref: "#/components/schemas/Doctor" } }
            reviews: { type: array, items: { $ref: "#/components/schemas/Review" } }
          required: [doctors, reviews]
...
    Doctor:
      type: object
      properties:
        id: { type: integer }
        hospitalId: { type: integer }
        name: { type: string }
        specialtyId: { type: integer }
        specialty: { type: string }
        qualifications: { type: string }
        yearsExperience: { type: integer }
        available: { type: boolean }
      required: [id, hospitalId, name, specialtyId, specialty, qualifications, yearsExperience, available]
...
    Review:
      type: object
      properties:
        id: { type: integer }
        hospitalId: { type: integer }
        hospitalName: { type: string }
        userId: { type: string }
        userName: { type: string }
        rating: { type: integer, minimum: 1, maximum: 5 }
        text: { type: string }
        status: { type: string, enum: [visible, hidden] }
        createdAt: { type: string, format: date-time }
        isDemo: { type: boolean }
      required: [id, hospitalId, hospitalName, userId, userName, rating, text, status, createdAt, isDemo]
    ReviewInput:
      type: object
      properties:
        rating: { type: integer, minimum: 1, maximum: 5 }
        text: { type: string, minLength: 1, maxLength: 1200 }
      required: [rating, text]
...
    EmergencyFacility:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        type: { type: string }
        address: { type: string }
        city: { type: string }
        phone: { type: string }
        category: { type: string }
        availabilityNote: { type: string }
        isDemo: { type: boolean }
      required: [id, name, type, address, city, phone, category, availabilityNote, isDemo]